delete: NotificationHandler를 삭제, SceneDelegate로 메서드 이동
코디네이터 패턴 리팩토링을 위해 NotificationHandler를 삭제
